+++ tests/machine/x/cs/avg_builtin.cs
@@ -6,6 +6,6 @@ class Program
 {
     static void Main()
     {
-        Console.WriteLine(Enumerable.Average(new List<long> { 1, 2, 3 }.Select(_tmp0 => Convert.ToDouble(_tmp0))));
+        Console.WriteLine(Enumerable.Average(new List<int> { 1, 2, 3 }.Select(_tmp0 => Convert.ToDouble(_tmp0))));
     }
 }
