+++ tests/machine/x/cs/break_continue.cs
@@ -5,7 +5,7 @@ class Program
 {
     static void Main()
     {
-        List<long> numbers = new List<long> { 1, 2, 3, 4, 5, 6, 7, 8, 9 };
+        List<int> numbers = new List<int> { 1, 2, 3, 4, 5, 6, 7, 8, 9 };
         foreach (var n in numbers)
         {
             if ((n % 2) == 0)
