+++ tests/machine/x/cs/map_literal_dynamic.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 
 class Program {
     static void Main() {
-        long x = 3;
-        long y = 4;
-        Dictionary<string, long> m = new Dictionary<string, long> { { "a", x }, { "b", y } };
+        int x = 3;
+        int y = 4;
+        Dictionary<string, int> m = new Dictionary<string, int> { { "a", x }, { "b", y } };
         Console.WriteLine(string.Join(" ", new [] { Convert.ToString(m["a"]), Convert.ToString(m["b"]) }));
     }
 }
