+++ tests/machine/x/cs/partial_application.cs
@@ -1,15 +1,12 @@
 using System;
 
-class Program
-{
-    static long add(long a, long b)
-    {
+class Program {
+    static long add(long a, long b) {
         return (a + b);
     }
-
-    static void Main()
-    {
-        long add5 = new Func<long, long>((long p0) => { return add(5, p0); });
+    
+    static void Main() {
+        int add5 = new Func<int, int>((int p0) => { return add(5, p0); });
         Console.WriteLine(add5(3));
     }
 }
